Use workflow nodes for workflow tests (#807)
